Weather condition modeling and prediction
This project leverage on core competence in data preprocessing and important feature selection techniques.
The concept of data balancing was adopted and the SMOTEENN -Synthetic Minority OverSampling Technique - Edited Nearest Neighbors was found to be efficient in handling imbalance dataset. Correcting the data imbalance provides opportunity for model performances to be optimized.